+++ temba/api/models.py
@@ -128,7 +128,6 @@ class Resthook(SmartModel):
 
     def add_subscriber(self, url, user):
         subscriber = self.subscribers.create(target_url=url, created_by=user, modified_by=user)
-        self.modified_on = timezone.now()
         self.modified_by = user
         self.save(update_fields=['modified_on', 'modified_by'])
         return subscriber
@@ -136,7 +135,6 @@ class Resthook(SmartModel):
     def remove_subscriber(self, url, user):
         now = timezone.now()
         self.subscribers.filter(target_url=url, is_active=True).update(is_active=False, modified_on=now, modified_by=user)
-        self.modified_on = now
         self.modified_by = user
         self.save(update_fields=['modified_on', 'modified_by'])
 
@@ -148,7 +146,6 @@ class Resthook(SmartModel):
         # then ourselves
         self.is_active = False
         self.modified_by = user
-        self.modified_on = timezone.now()
         self.save(update_fields=['is_active', 'modified_on', 'modified_by'])
 
     def as_select2(self):
@@ -172,11 +169,9 @@ class ResthookSubscriber(SmartModel):
     def release(self, user):
         self.is_active = False
         self.modified_by = user
-        self.modified_on = timezone.now()
         self.save(update_fields=['is_active', 'modified_on', 'modified_by'])
 
         # update our parent as well
-        self.resthook.modified_on = self.modified_on
         self.resthook.modified_by = user
         self.resthook.save(update_fields=['modified_on', 'modified_by'])